From: Chris Wilson <chris@aptivate.org> Date: Fri, 2 Dec 2011 20:38:26 +0000 (GMT)
I tried to use the tcpping probe but it didn't work for me (centos 5). It seems
that the tcptraceroute output format has changed or there is a problem with the
script's ability to parse options. It also doesn't report tcptraceroute errors
properly or have much in the way of debugging support. And it required
smokeping to be run as root.

I think I have fixed these issues, and posted updated versions here:

https://github.com/aptivate/network-scripts/blob/master/TCPPing.pm
https://github.com/aptivate/network-scripts/blob/master/tcpping

I hope that you will consider these for inclusion in your projects. Richard,
are you still maintaining tcpping? If not, is it worth merging this
functionality into the TCPPing module?

(integrated the TCPPing.pm for now -- tobi)