Fix unitialized value warnings when rrd returns undef. Add slave support (#42)
Got rid of the warning that happens when RRD returns 'undef' in the math on line 125.

Added support for slave rrd files using a cli option. If option exists all slaves are used. If a csv list of slaves is submitted with the option then only those slaves are used. Exact name match is required.

Added regex substitution for a '~' in the rrd name to make it readable as 'host from server' in morning email.

Added support for 'nomasterpoll' but only if processing for slaves. Having 'nomasterpoll' set for a host and no slaves configured for it is technically a configuration error, isn't it? If so, then this will show up as a 'down host' in the morning email this way.