FPing: create new parameter to force using IPv4 or IPv6 fixes#95
Starting with FPing 3.16, the IPv6 functionality was merged in the main
binary and fping now respects the system-wide setting for address
resolution. This means that when we add a target that has both an A
record and a AAAA record, the probe tests the IPv6 address by default.

This breaks the expectation that the FPing probe tests IPv4 and that the
FPing6 probe checks IPv6.

To restore this, we can now add a parameter to the probes to force the
protocol version to be 4 or 6.