AnotherDNS: Add expect_text and require_nxdomain target variables
expect_text allows testing for the presence of a string in the DNS
response record(s). This can be used to verify that an A record contains
the expected IP address, a PTR record reflects the expected hostname,
etc. If the query returns multiple records, any single match will pass
the test.

require_nxdomain toggles whether an NXDOMAIN response code should be
interpreted as success instead of failure. This reverses the normal
behavior of the probe. Example uses include testing a DNS firewall,
verifying that a mail server IP is not listed on a DNSBL, or other
scenarios where NXDOMAIN is desired.