FPing: disable use of -4/-6 if binary is too old (closes: #146)
It would seem as though some distribution are still shipping versions of
fping that are too old for the new -4 and -6 switches. Users of Centos 7
and RHEL 7 reported having trouble with the new "protocol" parameter
because of this: EPEL7 currently ships fping 3.10.

If an old binary is found when starting up our FPing probe, we'll
disable use of the new switches in order to maintain functionality to
users with older versions of fping.