Warn users if a probe binary exits with an error code (closes: #151)
When a probe fails to run at all, it currently fails silently and the
only method that users can use to determine that the binary is failing
is to strace the smokeping protocol.

Users can already enable debug output in order to gain some insight
about what is going wrong, but they need to know first that something is
wrong.

TraceroutePing was already checking the return code, but was only
outputting debug information about the binary erroring out. This was
changed to "carp" instead so that probe binary errors are visible in
logs.

Some discrepancies:

 * FPingContinuous doesn't show the command that failed when it does
   since the arguments are hidden away in the run_fping() function. This
   shouldn't be too great an issue, since the command is output in debug,
   so users can see what's happening when they enable debug.

 * IRTT already handles the return code but dies out when return code is
   non-zero. Maybe this should be changed to use "carp" like other probes?