Switch graphs over to SVG format (#451)
This fixes  a very long standing issue regarding the web pages on Hi-DPI
screens, as I can find messages from 2017. I filled myself #357 two
years ago, but I didn't managed until now to send a patch.

I've been running basically this diff for a couple of years, and it
works out of the box on my few installations, on high resolutions
screens. I'm not sure how to test on standard resolution screens, but
given that I didn't change anything related to resolutions, it should
work out of the box.

In #357 I mentioned having a configurable switch, but upon further
thinking, and the fact that switching to SVG is transparent, I don't
know if there's much gain from keeping both. There is one that might be
meaningful - PNG files, being binary, are "natively" compressed. SVG
files are not by default compressed, and since BREACH, most
installations disable compression for HTTPS. However, if this is a
concern, compression on SVG paths can be enabled in a server
configuration, as all the XML SVG files should not contain any tokens.
Happy to further discuss this concern.