Large whitespace cleanup
In my previous patch, my editor's setting automatically cleanup up
whitespace so it resulted in large diff, and I had to revert that.
Looking in the tree, there's widespread case of whitespace-at-eol or
double empty line at EOF.

I fully understand that this is mostly a style issue, and the current
situation doesn't break anything, so I won't mind at all if this patch
is rejected. However, in the case this just happened to be like this
accidentally, here is a cleanup.

I didn't touch the github workflows files as there the YAML cleaner
wants to do more stuff, I'll send that separately.